book/examples/fsl/requirements.txt
# data
mirdata
librosa

# training
torch
numpy
torchaudio
torchmetrics
pytorch-lightning

# display
torchvision
sklearn
umap-learn
pandas
plotly
kaleido
